Move ParamConverter API to jersey-core-common module
