+++ core-common/src/main/java/org/glassfish/jersey/internal/inject/PrimitiveMapper.java
@@ -13,8 +13,9 @@
  *
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  */
+// Portions Copyright [2018] [Payara Foundation and/or its affiliates]
 
-package org.glassfish.jersey.server.internal.inject;
+package org.glassfish.jersey.internal.inject;
 
 import java.util.Collections;
 import java.util.Map;
@@ -27,11 +28,11 @@ import java.util.WeakHashMap;
  * @author Paul Sandoz
  * @author Marek Potociar (marek.potociar at oracle.com)
  */
-final class PrimitiveMapper {
+public final class PrimitiveMapper {
 
-    static final Map<Class, Class> primitiveToClassMap =
+    public static final Map<Class, Class> primitiveToClassMap =
             getPrimitiveToClassMap();
-    static final Map<Class, Object> primitiveToDefaultValueMap =
+    public static final Map<Class, Object> primitiveToDefaultValueMap =
             getPrimitiveToDefaultValueMap();
 
     private static Map<Class, Class> getPrimitiveToClassMap() {
