+++ core-common/src/main/resources/org/glassfish/jersey/internal/localization.properties
@@ -13,6 +13,8 @@
 #
 # SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
 #
+# Portions Copyright [2018] [Payara Foundation and/or its affiliates]
+#
 
 
 # {0} - full classname
@@ -47,6 +49,7 @@ error.msg=WARNING: {0}
 error.newcookie.expires=NewCookie Expires header value ({0}) cannot be read.
 error.notfound.messagebodywriter=MessageBodyWriter not found for media type={0}, type={1}, genericType={2}.
 error.notfound.messagebodyreader=MessageBodyReader not found for media type={0}, type={1}, genericType={2}.
+error.parameter.invalid.char.value=Value "{0}" is not a character.
 error.parsing.entity.tag=Error parsing entity tag: {0}
 error.provider.constrainedTo.wrong.package=A registered provider {0} is constrained (via @ConstrainedTo) to {1} runtime but does not implement any provider interface usable in the runtime.
 error.provider.constrainedTo.wrong.runtime=A provider {0} registered in {2} runtime is constrained (via @ConstrainedTo) to {1} runtime.
@@ -107,6 +110,7 @@ message.content.buffering.failed=Failed to buffer the message content input stre
 message.content.input.stream.close.failed=Error closing message content input stream.
 message.content.buffer.reset.failed=Error resetting the buffered message content input stream.
 method.not.getter.nor.setter=Method is neither getter nor setter.
+method.parameter.cannot.be.null=Method parameter "{0}" cannot be null.
 multiple.matching.constructors.found=Found {0} constructors with {1} parameters in {2} class. Selecting the first found constructor: {3}
 new.cookie.is.null=New cookie is null.
 no.container.available=No container available.
